pyproject.toml
[tool.poetry]
name = "sql-unit-test"
version = "0.1.0"
description = "A CLI tool for executing sql unit tests."
authors = ["cohenj20 <cohenj20@sacredheart.edu>"]
readme = "README.md"
license = "MIT"
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.10"
colorama = "^0.4.6"
click = "^8.1.7"
pandas = "^2.2.1"
pyodbc = "^5.1.0"
sqlalchemy = "2.0.28"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
sql-unit-test = 'sql_unit_test.main:main'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
